FINALLY DONE WITH DAY 11

+++ src/main.rs
@@ -3,6 +3,8 @@ use std::error::Error;
 use std::fs;
 mod day1;
 mod day10;
+mod day11;
+mod day12;
 mod day2;
 mod day3;
 mod day4;
@@ -13,7 +15,7 @@ mod day8;
 mod day9;
 fn set_up_day(aoc_cookie: &str, day: &str) -> Result<(), Box<dyn Error>> {
     println!("setting up day {}...", day);
-    let url = format!("https://adventofcode.com/day{}/input", day);
+    let url = format!("https://adventofcode.com/2022/day/{}/input", day);
     let dir = format!("src/day{}", day);
     let client = reqwest::blocking::Client::new();
     let res = client
@@ -22,9 +24,21 @@ fn set_up_day(aoc_cookie: &str, day: &str) -> Result<(), Box<dyn Error>> {
         .send()?;
     if res.status().is_success() {
         let input = res.text()?;
-        fs::create_dir(dir.as_str())?;
-        fs::write(format!("{}/input.txt", dir), input)?;
-        println!("done. don't forget to add the module and module function to main.rs");
+        let src_entries = fs::read_dir("src")?.collect::<Vec<_>>();
+        let existing = src_entries
+            .iter()
+            .filter(|e| match e {
+                Ok(name) => name.file_name().to_str().unwrap().contains(day),
+                Err(_) => false,
+            })
+            .collect::<Vec<_>>();
+        if existing.len() == 0 {
+            fs::create_dir(dir.as_str())?;
+            fs::write(format!("{}/input.txt", dir), input)?;
+            println!("done. don't forget to add the module and module function to main.rs");
+        } else {
+            println!("dir exists so you must remove");
+        }
     } else {
         println!("page or day does not exist :(");
     }
@@ -58,6 +72,11 @@ fn main() -> Result<(), Box<dyn Error>> {
         "8" => day8::day8(),
         "9" => day9::day9(),
         "10" => day10::day10(),
+        "11" => {
+            day11::day11_p1();
+            day11::day11_p2();
+        }
+        "12" => {}
         _ => set_up_day(aoc_cookie.as_str(), day.as_str())?,
     }
     Ok(())
